move trigger file
